(REF) Split `AllFlowsTest` into smaller, thematic tests
